Bug #1347: quota update accepts float values (cherry picked from commit c94bfcab5e819b0f494a4f7c31ad3b7dd8ed1a22)

+++ src/um/QuotaDatastore.cc
@@ -58,7 +58,7 @@ bool QuotaDatastore::check(Template * tmpl,  string& error)
 
 void QuotaDatastore::del(Template * tmpl)
 {
-    map<string, int> ds_request;
+    map<string, float> ds_request;
 
     string ds_id;
     int    size;
